\name{fACD-package}
\alias{fACD-package}
\alias{fACD}
\docType{package}
\title{Package for Estimation and Simulation of ACD Models}
\description{This package can estimate and simulate ACD (Autoregressive Conditional Duration)
models, with support for the ACD, log, BC and EX specifications, also including the choice for the exponential
or the weibull function as the distribution assumption.}
\details{
\tabular{ll}{
Package: \tab fACD\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-05-05\cr
License: \tab GPL version 2 or newer  \cr
LazyLoad: \tab yes\cr
}}
\author{Marcelo Perlin - ICMA/UK <\email{marceloperlin@gmail.com}>}
\keyword{models}